requirements.txt
python-dotenv
twitchio
tinydb